Extract code from js utilities
