+++ src/language-js/comments/handle-comments.js
@@ -9,8 +9,10 @@ import hasNewline from "../../utilities/has-newline.js";
 import hasNewlineInRange from "../../utilities/has-newline-in-range.js";
 import isNonEmptyArray from "../../utilities/is-non-empty-array.js";
 import { locEnd, locStart } from "../loc.js";
+import { getCallArguments } from "../utilities/call-arguments.js";
 import { isInsideCallOrNewExpressionParentheses } from "../utilities/call-or-new-expression-parentheses.js";
 import { createTypeCheckFunction } from "../utilities/create-type-check-function.js";
+import { getFunctionParameters } from "../utilities/function-parameters.js";
 import getTextWithoutComments from "../utilities/get-text-without-comments.js";
 import isBlockComment from "../utilities/is-block-comment.js";
 import isLineComment from "../utilities/is-line-comment.js";
@@ -25,12 +27,7 @@ import {
   isMemberExpression,
   isUnionType,
 } from "../utilities/node-types.js";
-import {
-  getCallArguments,
-  getFunctionParameters,
-  isMethod,
-  isObjectProperty,
-} from "../utilities/utilities.js";
+import { isMethod, isObjectProperty } from "../utilities/utilities.js";
 
 /**
 @import {Node, Comment, NodeMap} from "../types/estree.js";
