+++ src/language-js/print/call-arguments.js
@@ -10,7 +10,15 @@ import {
   softline,
   willBreak,
 } from "../../document/index.js";
+import {
+  getCallArguments,
+  getCallArgumentSelector,
+  iterateCallArgumentsPath,
+} from "../utilities/call-arguments.js";
 import { CommentCheckFlags, hasComment } from "../utilities/comments.js";
+import { getFunctionParameters } from "../utilities/function-parameters.js";
+import { isFunctionCompositionArguments } from "../utilities/is-function-composition-arguments.js";
+import { isSimpleCallArgument } from "../utilities/is-simple-call-argument.js";
 import {
   isArrayExpression,
   isBinaryCastExpression,
@@ -22,19 +30,13 @@ import {
   isRegExpLiteral,
   isStringLiteral,
 } from "../utilities/node-types.js";
+import { stripChainElementWrappers } from "../utilities/strip-chain-element-wrappers.js";
 import {
-  getCallArguments,
-  getCallArgumentSelector,
-  getFunctionParameters,
-  isFunctionCompositionArgs,
   isLongCurriedCallExpression,
   isNextLineEmpty,
   isObjectProperty,
-  isSimpleCallArgument,
   isSimpleType,
-  iterateCallArgumentsPath,
   shouldPrintComma,
-  stripChainElementWrappers,
 } from "../utilities/utilities.js";
 import { isConciselyPrintedArray } from "./array.js";
 import { printDanglingCommentsInList } from "./miscellaneous.js";
@@ -109,7 +111,7 @@ function printCallArguments(path, options, print) {
 
   if (
     anyArgEmptyLine ||
-    (path.parent.type !== "Decorator" && isFunctionCompositionArgs(args))
+    (path.parent.type !== "Decorator" && isFunctionCompositionArguments(args))
   ) {
     return allArgsBrokenOut();
   }
